requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
Pillow>=10.0.0
numpy>=1.24.0
pytest>=7.4.0
